youtube/requirements.txt
google-api-python-client==2.147.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0
requests==2.32.3
python-dotenv==1.0.1
